# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ToyFinancial dealer registration portal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: public
    description: Public registration operations
  - name: admin
    description: Authenticated administration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [public]
      summary: List available dealer programs
      responses:
        "200":
          description: Available programs and provinces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductCatalog"
  /registrations:
    post:
      operationId: createRegistration
      tags: [public]
      summary: Submit a dealer registration request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationReceipt"
        "400":
          description: Invalid registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in to the admin portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Signed-in administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid sign-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Sign out of the admin portal
      responses:
        "204":
          description: Signed out
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Get the current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/metrics:
    get:
      operationId: getAdminMetrics
      tags: [admin]
      summary: Get registration pipeline metrics
      responses:
        "200":
          description: Pipeline metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMetrics"
        "401":
          description: Admin sign-in required
  /admin/dealers:
    get:
      operationId: listDealers
      tags: [admin]
      summary: List dealer registrations
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Dealer registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dealer"
        "401":
          description: Admin sign-in required
  /admin/dealers/{id}:
    get:
      operationId: getDealer
      tags: [admin]
      summary: Get one dealer registration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dealer registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dealer"
        "404":
          description: Dealer not found
        "401":
          description: Admin sign-in required
    patch:
      operationId: updateDealer
      tags: [admin]
      summary: Update a dealer registration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealerUpdate"
      responses:
        "200":
          description: Updated dealer registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dealer"
        "400":
          description: Invalid update
        "404":
          description: Dealer not found
        "401":
          description: Admin sign-in required
  /admin/documents:
    get:
      operationId: listDocuments
      tags: [admin]
      summary: List starter-pack documents
      responses:
        "200":
          description: Starter-pack documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          description: Admin sign-in required
    post:
      operationId: createDocument
      tags: [admin]
      summary: Add a starter-pack document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid document
        "401":
          description: Admin sign-in required
  /admin/documents/{id}:
    patch:
      operationId: updateDocument
      tags: [admin]
      summary: Update a starter-pack document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found
        "401":
          description: Admin sign-in required
    delete:
      operationId: deleteDocument
      tags: [admin]
      summary: Delete a starter-pack document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Document not found
        "401":
          description: Admin sign-in required
  /admin/outbox:
    get:
      operationId: listOutbox
      tags: [admin]
      summary: List recent email outbox items
      responses:
        "200":
          description: Recent email outbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutboxItem"
        "401":
          description: Admin sign-in required
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ProductCatalog:
      type: object
      properties:
        products:
          type: array
          items:
            type: string
        provinces:
          type: array
          items:
            type: string
      required: [products, provinces]
    RegistrationInput:
      type: object
      properties:
        dealerName: { type: string, minLength: 1, maxLength: 200 }
        contactName: { type: string, minLength: 1, maxLength: 200 }
        title: { type: string, maxLength: 200 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 1, maxLength: 50 }
        city: { type: string, minLength: 1, maxLength: 100 }
        province: { type: string, minLength: 1, maxLength: 100 }
        website: { type: string, maxLength: 500 }
        locations: { type: string, maxLength: 500 }
        products:
          type: array
          minItems: 1
          items: { type: string }
        categories:
          type: array
          items: { type: string }
        launchTiming: { type: string, maxLength: 100 }
        currentProvider: { type: string, maxLength: 200 }
        valueNotes: { type: string, maxLength: 2000 }
        followUpMethod: { type: string, maxLength: 100 }
        meetingInterest: { type: string, maxLength: 100 }
        consent:
          type: boolean
        materialsRequested:
          type: boolean
      required:
        - dealerName
        - contactName
        - email
        - phone
        - city
        - province
        - products
        - consent
        - materialsRequested
    RegistrationReceipt:
      type: object
      properties:
        id: { type: integer }
        message: { type: string }
      required: [id, message]
    AdminLoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    AdminSession:
      type: object
      properties:
        authenticated: { type: boolean }
        email: { type: string }
      required: [authenticated]
    AdminMetrics:
      type: object
      properties:
        total: { type: integer }
        newLeads: { type: integer }
        live: { type: integer }
        fulfillmentOpen: { type: integer }
      required: [total, newLeads, live, fulfillmentOpen]
    Dealer:
      type: object
      properties:
        id: { type: integer }
        dealerName: { type: string }
        contactName: { type: string }
        title: { type: string }
        email: { type: string }
        phone: { type: string }
        city: { type: string }
        province: { type: string }
        website: { type: string }
        locations: { type: string }
        products:
          type: array
          items: { type: string }
        categories:
          type: array
          items: { type: string }
        launchTiming: { type: string }
        currentProvider: { type: string }
        valueNotes: { type: string }
        followUpMethod: { type: string }
        meetingInterest: { type: string }
        consentAt: { type: string, format: date-time }
        leadStatus: { type: string }
        assignedTo: { type: string }
        lastContactedAt: { type: ["string", "null"], format: date-time }
        notes: { type: string }
        fulfillmentStatus: { type: string }
        fulfillmentNotes: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - dealerName
        - contactName
        - title
        - email
        - phone
        - city
        - province
        - website
        - locations
        - products
        - categories
        - launchTiming
        - currentProvider
        - valueNotes
        - followUpMethod
        - meetingInterest
        - consentAt
        - leadStatus
        - assignedTo
        - lastContactedAt
        - notes
        - fulfillmentStatus
        - fulfillmentNotes
        - createdAt
        - updatedAt
    DealerUpdate:
      type: object
      properties:
        dealerName: { type: string, minLength: 1, maxLength: 200 }
        contactName: { type: string, minLength: 1, maxLength: 200 }
        title: { type: string, maxLength: 200 }
        email: { type: string, format: email }
        phone: { type: string, maxLength: 50 }
        city: { type: string, maxLength: 100 }
        province: { type: string, maxLength: 100 }
        website: { type: string, maxLength: 500 }
        locations: { type: string, maxLength: 500 }
        products:
          type: array
          items: { type: string }
        categories:
          type: array
          items: { type: string }
        launchTiming: { type: string, maxLength: 100 }
        currentProvider: { type: string, maxLength: 200 }
        valueNotes: { type: string, maxLength: 2000 }
        followUpMethod: { type: string, maxLength: 100 }
        meetingInterest: { type: string, maxLength: 100 }
        leadStatus: { type: string }
        assignedTo: { type: string, maxLength: 200 }
        lastContactedAt: { type: ["string", "null"], format: date-time }
        notes: { type: string, maxLength: 2000 }
        fulfillmentStatus: { type: string }
        fulfillmentNotes: { type: string, maxLength: 2000 }
    Document:
      type: object
      properties:
        id: { type: integer }
        product: { type: string }
        province: { type: string }
        kind: { type: string }
        title: { type: string }
        url: { type: string, format: uri }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, product, province, kind, title, url, active, createdAt, updatedAt]
    DocumentInput:
      type: object
      properties:
        product: { type: string }
        province: { type: string }
        kind: { type: string, minLength: 1 }
        title: { type: string, minLength: 1, maxLength: 200 }
        url: { type: string, format: uri }
        active: { type: boolean }
      required: [product, kind, title, url, active]
    OutboxItem:
      type: object
      properties:
        id: { type: integer }
        recipient: { type: string }
        subject: { type: string }
        type: { type: string }
        deliveryStatus: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, recipient, subject, type, deliveryStatus, createdAt]