# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ToyFinancial dealer registration portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  authenticated: boolean;
  email?: string;
}