# lib/api-zod/src/generated/types/dealer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ToyFinancial dealer registration portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Dealer {
  id: number;
  dealerName: string;
  contactName: string;
  title: string;
  email: string;
  phone: string;
  city: string;
  province: string;
  website: string;
  locations: string;
  products: string[];
  categories: string[];
  launchTiming: string;
  currentProvider: string;
  valueNotes: string;
  followUpMethod: string;
  meetingInterest: string;
  consentAt: Date;
  leadStatus: string;
  assignedTo: string;
  /** @nullable */
  lastContactedAt: Date | null;
  notes: string;
  fulfillmentStatus: string;
  fulfillmentNotes: string;
  createdAt: Date;
  updatedAt: Date;
}