# lib/api-zod/src/generated/types/dealerUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ToyFinancial dealer registration portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DealerUpdate {
  /**
     * @minLength 1
     * @maxLength 200
     */
  dealerName?: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  contactName?: string;
  /** @maxLength 200 */
  title?: string;
  email?: string;
  /** @maxLength 50 */
  phone?: string;
  /** @maxLength 100 */
  city?: string;
  /** @maxLength 100 */
  province?: string;
  /** @maxLength 500 */
  website?: string;
  /** @maxLength 500 */
  locations?: string;
  products?: string[];
  categories?: string[];
  /** @maxLength 100 */
  launchTiming?: string;
  /** @maxLength 200 */
  currentProvider?: string;
  /** @maxLength 2000 */
  valueNotes?: string;
  /** @maxLength 100 */
  followUpMethod?: string;
  /** @maxLength 100 */
  meetingInterest?: string;
  leadStatus?: string;
  /** @maxLength 200 */
  assignedTo?: string;
  /** @nullable */
  lastContactedAt?: Date | null;
  /** @maxLength 2000 */
  notes?: string;
  fulfillmentStatus?: string;
  /** @maxLength 2000 */
  fulfillmentNotes?: string;
}